Handling NULL values for group day
A group's day can be a NULL value when "Various Days" is selected as the option within ChurchSuite. Previously, we weren't handling this scenario, which resulted in a JS error. This updates the logic so that, when filtering groups, we consider a group to match the day search filter if the group meets on "Various Days"/NULL or we have an actual match.