Updating signup_enabled to lean on the new api array instead of looking for the url.
We’ve decided to go with our rudimentary HTTP_REFERER blocking of embed sign-up if it isn’t enabled so we always send the event page URL across. As we were relying on the presence of a link to determine whether signup was enabled or not we’ve had to provide a boolean prop in the API array which reflects signup enabled and also embed signup enabled.